requirements.txt
aiohttp==3.6.2
async-timeout==3.0.1
attrs==19.3.0
cachetools==4.1.0
certifi==2020.4.5.1
chardet==3.0.4
discord.py==1.3.3
google-api-core==1.17.0
google-api-python-client==1.8.4
google-auth==1.16.0
google-auth-httplib2==0.0.3
googleapis-common-protos==1.51.0
httplib2==0.18.1
idna==2.9
multidict==4.7.6
protobuf==3.12.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pymongo==3.10.1
python-dotenv==0.13.0
pytz==2020.1
requests==2.23.0
rsa==4.0
six==1.15.0
uritemplate==3.0.1
urllib3==1.25.9
websockets==8.1
yarl==1.4.2
